# schemas/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "SWEN Bank API",
    "description": "API Documentation For SWEN Bank",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3030",
  "basePath": "/",
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/Receivers": {
      "post": {
        "tags": [
          "Receivers"
        ],
        "summary": "Add an account as a receiver in the receiver list. Only User role can access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the account to add to the receiver list",
            "required": true,
            "schema": {
              "properties": {
                "reminiscent_name": {
                  "type": "string"
                },
                "account_number": {
                  "type": "string"
                },
                "bank_id": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully create a new receiver",
            "schema": {
              "properties": {
                "id": {
                  "type": "integer"
                },
                "reminiscent_name": {
                  "type": "string"
                },
                "user_id": {
                  "type": "integer"
                },
                "account_number": {
                  "type": "string"
                },
                "bank_id": {
                  "type": "integer"
                }
              }
            }
          },
          "200": {
            "description": "Cannot find the account",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Receivers"
        ],
        "summary": "Get all receiver in the receiver list. Only User role can access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Get the receiver list",
            "schema": {
              "properties": {
                "id": {
                  "type": "integer"
                },
                "reminiscent_name": {
                  "type": "string"
                },
                "user_id": {
                  "type": "integer"
                },
                "account_number": {
                  "type": "string"
                },
                "bank_id": {
                  "type": "integer"
                }
              }
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/Receivers/{receiverID}": {
      "patch": {
        "tags": [
          "Receivers"
        ],
        "summary": "Update a receiver in the receiver list. Only User role can access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "receiverID",
            "in": "path",
            "required": true,
            "description": "ID of a receiver",
            "type": "integer"
          },
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the receiver information to be updated",
            "required": true,
            "schema": {
              "properties": {
                "reminiscent_name": {
                  "type": "string"
                },
                "account_number": {
                  "type": "string"
                },
                "bank_id": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cannot find the receiver's account",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "201": {
            "description": "Update receiver successfully",
            "schema": {
              "properties": {
                "id": {
                  "type": "integer"
                },
                "reminiscent_name": {
                  "type": "string"
                },
                "user_id": {
                  "type": "integer"
                },
                "account_number": {
                  "type": "string"
                },
                "bank_id": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Receivers"
        ],
        "summary": "Delete a receiver. Only User role can access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "receiverID",
            "in": "path",
            "required": true,
            "description": "ID of a receiver",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Cannot find the receiver's account",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "201": {
            "description": "Successfully delete the receiver",
            "schema": {
              "$ref": "#/definitions/successResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/Transactions": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Make a transaction. Must be login as User role to access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the transaction information to proceed",
            "required": true,
            "schema": {
              "properties": {
                "source_account_uuid": {
                  "type": "string"
                },
                "destination_account_uuid": {
                  "type": "string"
                },
                "amount": {
                  "type": "integer"
                },
                "note": {
                  "type": "string"
                },
                "fee_is_paid_by_receiver": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cannot find soure or receiver account",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "201": {
            "description": "Transaction transfer ID",
            "schema": {
              "properties": {
                "id": {
                  "type": "integer"
                }
              }
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/Transactions/VerifyOTP": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Verify OTP code to confirm the transaction. Must be login as User role to access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "Request Body",
            "in": "body",
            "description": "Provide the OTP code to confirm the transaction",
            "required": true,
            "schema": {
              "properties": {
                "otp": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invalid OTP or bank account not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "201": {
            "description": "Transaction has been confirmed",
            "schema": {
              "$ref": "#/definitions/successResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/Transactions/{Transaction_ID}/ResendOTP": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Re-send another OTP code to confirm the transaction. Must be login as User role to access this API",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Provide the access token to access the API",
            "required": true
          },
          {
            "name": "Transaction_ID",
            "in": "path",
            "required": true,
            "description": "ID of a transaction",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Failed to send OTP code",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "201": {
            "description": "New OTP code is sent",
            "schema": {
              "$ref": "#/definitions/successResponse"
            }
          },
          "403": {
            "description": "Do not have permission to access the API",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "500": {
            "description": "Internal server error!",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "errorResponse": {
      "properties": {
        "message": {
          "type": "string",
          "description": "The error message"
        }
      }
    },
    "successResponse": {
      "properties": {
        "message": {
          "type": "string",
          "description": "The success message"
        }
      }
    }
  }
}